Moved interfaces to separate file
