+++ src/app.ts
@@ -12,40 +12,6 @@ import { rateLimiter } from './utils/rateLimiter'
 export const app = express()
 export const port = 42069
 
-export const startingFundsDefault = 1000
-export const smallBlindDefault = 100
-export interface PlayerInfo {
-  nickname: string
-  playerHash: string
-}
-export interface InternalPlayerInfo {
-  token: string
-  card1: string
-  card2: string
-}
-
-export interface PlayerGameInfo {
-  playerHash: string
-  turn: number
-}
-export interface GameSettings {
-  smallBlind: number
-  startingFunds: number
-  players: PlayerInfo[]
-  gameMasterHash: string
-}
-
-export interface NewGameInfo {
-  gameId: number
-  startingFunds: number
-  smallBlind: number
-}
-
-export interface StartingGameInfo {
-  players: PlayerGameInfo[]
-  cards: string[]
-}
-
 const errorHandling = (error, req, res, next) => {
   if (isCelebrateError(error)) {
     return res.status(400).send({
