+++ src/routes/joinGame.ts
@@ -1,7 +1,7 @@
 import { getClient } from '../utils/databaseConnection'
 import { celebrate, Joi, Segments } from 'celebrate'
 import sha256 from 'crypto-js/sha256'
-import { type GameSettings } from '../app'
+import type { GameSettings } from '../utils/types'
 import { sendFirebaseMessage, verifyFCMToken } from '../utils/firebase'
 import express, { type Router } from 'express'
 import { rateLimiter } from '../utils/rateLimiter'
