+++ src/routes/startGame.ts
@@ -2,7 +2,7 @@ import { getClient } from '../utils/databaseConnection'
 import { rateLimiter } from '../utils/rateLimiter'
 import { celebrate, Joi, Segments } from 'celebrate'
 import { sendFirebaseMessage, verifyFCMToken } from '../utils/firebase'
-import { type StartingGameInfo, type InternalPlayerInfo } from '../app'
+import type { StartingGameInfo, InternalPlayerInfo } from '../utils/types'
 import { shuffleArray, fullCardDeck } from '../utils/randomise'
 import sha256 from 'crypto-js/sha256'
 import { type Client } from 'pg'
